+++ day9/day9_test.go
@@ -15,12 +15,12 @@ func TestPart1(t *testing.T) {
 	output := &intcomputer.OutputRecorder{}
 	c := intcomputer.NewIntComputer(realProgram, output.HandleOutput, intcomputer.SimpleInputter(1))
 	c.RunOperations()
-	assert.Equal(t, []int{3906448201}, output.Outputs)
+	assert.Equal(t, []int64{3906448201}, output.Outputs)
 }
 
 func TestPart2(t *testing.T) {
 	output := &intcomputer.OutputRecorder{}
 	c := intcomputer.NewIntComputer(realProgram, output.HandleOutput, intcomputer.SimpleInputter(2))
 	c.RunOperations()
-	assert.Equal(t, []int{59785}, output.Outputs)
+	assert.Equal(t, []int64{59785}, output.Outputs)
 }
